Fixed #198, bump oxalis to 5.5.0 and other dependecies upgrade
