+++ pom.xml
@@ -27,7 +27,7 @@
     <parent>
         <groupId>network.oxalis</groupId>
         <artifactId>oxalis</artifactId>
-        <version>5.4.0</version>
+        <version>5.5.0</version>
     </parent>
 
     <artifactId>oxalis-as4</artifactId>
@@ -87,23 +87,24 @@
         <connection>scm:git:https://github.com/OxalisCommunity/Oxalis-AS4.git</connection>
         <developerConnection>scm:git:git@github.com:OxalisCommunity/Oxalis-AS4.git</developerConnection>
         <url>https://github.com/OxalisCommunity/Oxalis-AS4</url>
-      <tag>HEAD</tag>
+        <tag>HEAD</tag>
     </scm>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
-
-        <peppol.sbdh.version>1.0.0</peppol.sbdh.version>
-
-        <cxf.version>3.3.8</cxf.version>
-        <wss4j.version>2.2.7</wss4j.version>
+        <peppol.sbdh.version>1.1.0</peppol.sbdh.version>
+        <cxf.version>3.5.5</cxf.version>
+        <wss4j.version>2.4.1</wss4j.version>
+        <jakarta.xml.bind-api.version>2.3.3</jakarta.xml.bind-api.version>
+        <jakarta.activation-api.version>1.2.1</jakarta.activation-api.version>
+        <bouncycastle.version>1.70</bouncycastle.version>
+        <commons-io.version>2.11.0</commons-io.version>
         <neethi.version>3.2.0</neethi.version>
-        <soap.version>1.4.0</soap.version>
-
+        <junit.version>4.13.2</junit.version>
         <powermock.version>2.0.0</powermock.version>
-        <mockito.version>2.10.0</mockito.version>
+        <mockito.version>2.23.4</mockito.version>
     </properties>
 
     <dependencies>
@@ -150,8 +151,8 @@
             <version>${cxf.version}</version>
             <exclusions>
                 <exclusion>
-                    <groupId>com.fasterxml.woodstox</groupId>
-                    <artifactId>woodstox-core</artifactId>
+                    <groupId>com.sun.activation</groupId>
+                    <artifactId>jakarta.activation</artifactId>
                 </exclusion>
             </exclusions>
         </dependency>
@@ -204,6 +205,10 @@
                     <groupId>org.apache.geronimo.javamail</groupId>
                     <artifactId>geronimo-javamail_1.4_mail</artifactId>
                 </exclusion>
+                <exclusion>
+                    <groupId>org.ehcache</groupId>
+                    <artifactId>ehcache</artifactId>
+                </exclusion>
             </exclusions>
         </dependency>
 
@@ -232,6 +237,22 @@
                 <exclusion>
                     <groupId>org.opensaml</groupId>
                     <artifactId>opensaml-xacml-saml-impl</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>com.google.guava</groupId>
+                    <artifactId>guava</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>com.fasterxml.woodstox</groupId>
+                    <artifactId>woodstox-core</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>org.ehcache</groupId>
+                    <artifactId>ehcache</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>org.bouncycastle</groupId>
+                    <artifactId>bcprov-jdk15on</artifactId>
                 </exclusion>
             </exclusions>
         </dependency>
@@ -272,24 +293,50 @@
                     <groupId>org.ehcache</groupId>
                     <artifactId>ehcache</artifactId>
                 </exclusion>
+                <exclusion>
+                    <groupId>org.bouncycastle</groupId>
+                    <artifactId>bcprov-jdk15on</artifactId>
+                </exclusion>
             </exclusions>
         </dependency>
         <dependency>
             <groupId>org.apache.wss4j</groupId>
             <artifactId>wss4j-policy</artifactId>
             <version>${wss4j.version}</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.apache.neethi</groupId>
+                    <artifactId>neethi</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
+            <groupId>jakarta.xml.soap</groupId>
+            <artifactId>jakarta.xml.soap-api</artifactId>
+            <version>1.4.2</version>
+        </dependency>
+        <dependency>
+            <groupId>com.sun.xml.messaging.saaj</groupId>
+            <artifactId>saaj-impl</artifactId>
+            <version>1.5.3</version>
+            <exclusions>
+                <exclusion>
+                    <groupId>com.sun.activation</groupId>
+                    <artifactId>jakarta.activation</artifactId>
+                </exclusion>
+            </exclusions>
         </dependency>
 
         <!-- Jakarta -->
         <dependency>
             <groupId>jakarta.xml.bind</groupId>
             <artifactId>jakarta.xml.bind-api</artifactId>
-            <version>2.3.2</version>
+            <version>${jakarta.xml.bind-api.version}</version>
         </dependency>
         <dependency>
             <groupId>jakarta.activation</groupId>
             <artifactId>jakarta.activation-api</artifactId>
-            <version>1.2.1</version>
+            <version>${jakarta.activation.version}</version>
         </dependency>
 
         <!-- Bouncycastle -->
@@ -319,7 +366,7 @@
         <dependency>
             <groupId>commons-io</groupId>
             <artifactId>commons-io</artifactId>
-            <version>2.7</version>
+            <version>${commons-io.version}</version>
         </dependency>
         <dependency>
             <groupId>commons-codec</groupId>
@@ -352,12 +399,6 @@
             <groupId>org.apache.neethi</groupId>
             <artifactId>neethi</artifactId>
             <version>${neethi.version}</version>
-            <exclusions>
-                <exclusion>
-                    <artifactId>woodstox-core-asl</artifactId>
-                    <groupId>org.codehaus.woodstox</groupId>
-                </exclusion>
-            </exclusions>
         </dependency>
 
         <dependency>
@@ -369,7 +410,7 @@
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-            <version>1.18.12</version>
+            <version>${lombok.version}</version>
             <scope>provided</scope>
         </dependency>
 
@@ -387,7 +428,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.13.1</version>
+            <version>${junit.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -537,6 +578,7 @@
                             <failOnWarning>true</failOnWarning>
                             <outputXML>true</outputXML>
                             <usedDependencies>
+                                <usedDependency>com.sun.xml.messaging.saaj:saaj-impl</usedDependency>
                                 <usedDependency>commons-codec:commons-codec</usedDependency>
                                 <usedDependency>org.kohsuke.metainf-services:metainf-services</usedDependency>
                             </usedDependencies>
