+++ src/test/java/network/oxalis/as4/common/DefaultMessageIdGeneratorTest.java
@@ -2,10 +2,10 @@ package network.oxalis.as4.common;
 
 import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
+import org.testng.annotations.Test;
 
 import java.util.UUID;
